third_party/blink/web_tests/external/wpt/css/css-lists/parsing/list-style-valid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['list-style'] = "disc outside none" should set the property value
  assert_equals: serialization should be canonical expected "outside" but got "outside none disc"
[FAIL] e.style['list-style'] = "inside disc" should set the property value
  assert_equals: serialization should be canonical expected "inside" but got "inside disc"
[FAIL] e.style['list-style'] = "inside none none" should set the property value
  assert_equals: serialization should be canonical expected "inside none" but got "inside none none"
[FAIL] e.style['list-style'] = "none inside none" should set the property value
  assert_equals: serialization should be canonical expected "inside none" but got "inside none none"
[FAIL] e.style['list-style'] = "none none inside" should set the property value
  assert_equals: serialization should be canonical expected "inside none" but got "inside none none"
[FAIL] e.style['list-style'] = "url(\\"https://example.com/\\") disc" should set the property value
  assert_equals: serialization should be canonical expected "url(\\"https://example.com/\\")" but got "url(\\"https://example.com/\\") disc"
[FAIL] e.style['list-style'] = "url(\\"https://example.com/\\") disc outside" should set the property value
  assert_equals: serialization should be canonical expected "url(\\"https://example.com/\\")" but got "outside url(\\"https://example.com/\\") disc"
[FAIL] e.style['list-style'] = "disc radial-gradient(circle, #006, #00a 90%, #0000af 100%,white 100%) inside" should set the property value
  assert_equals: serialization should be canonical expected "inside radial-gradient(circle, rgb(0, 0, 102), rgb(0, 0, 170) 90%, rgb(0, 0, 175) 100%, white 100%)" but got "inside radial-gradient(circle, rgb(0, 0, 102), rgb(0, 0, 170) 90%, rgb(0, 0, 175) 100%, white 100%) disc"
Harness: the test ran to completion.
